+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/EmployeeEditPanel.java
@@ -26,6 +26,7 @@ import javax.swing.JPanel;
 import static is.codion.demos.chinook.domain.api.Chinook.Employee;
 import static is.codion.swing.common.ui.component.Components.flexibleGridLayoutPanel;
 import static is.codion.swing.common.ui.component.Components.gridLayoutPanel;
+import static is.codion.swing.common.ui.key.TransferFocusOnEnter.BACKWARD;
 import static is.codion.swing.common.ui.layout.Layouts.flexibleGridLayout;
 
 public final class EmployeeEditPanel extends EntityEditPanel {
@@ -66,9 +67,10 @@ public final class EmployeeEditPanel extends EntityEditPanel {
 						.columns(12);
 		createComboBox(Employee.REPORTSTO_FK)
 						.preferredWidth(120)
-						// Disable transfer focus on Enter, so the Enter key triggers
-						// the default dialog button, when inserting and updating
-						.transferFocusOnEnter(false);
+						// Only transfer focus backward on Enter, this way
+						// the Enter key without any modifiers will trigger
+						// the default dialog button, for inserting and updating
+						.transferFocusOnEnter(BACKWARD);
 
 		JPanel firstLastNamePanel = gridLayoutPanel(1, 2)
 						.add(createInputPanel(Employee.FIRSTNAME))