+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -32,6 +32,7 @@ import is.codion.swing.framework.ui.EntityEditPanel;
 
 import javax.swing.AbstractButton;
 import javax.swing.Action;
+import javax.swing.JComponent;
 import javax.swing.JPanel;
 import java.awt.BorderLayout;
 import java.awt.event.FocusAdapter;
@@ -233,9 +234,9 @@ public final class EntityComboBoxPanel extends JPanel {
 		}
 
 		@Override
-		protected void enableTransferFocusOnEnter(EntityComboBoxPanel component) {
-			TransferFocusOnEnter.enable(component.comboBox());
-			component.buttons.forEach(TransferFocusOnEnter::enable);
+		protected void enableTransferFocusOnEnter(EntityComboBoxPanel component, TransferFocusOnEnter transferFocusOnEnter) {
+			transferFocusOnEnter.enable(component.comboBox());
+			transferFocusOnEnter.enable(component.buttons.toArray(new JComponent[0]));
 		}
 
 		@Override